requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
boto3==1.20.31
botocore==1.23.31
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-appconf==1.0.5
django-cors-headers==3.10.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
mccabe==0.6.1
msgpack==1.0.3
Pillow==8.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
requests==2.27.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
